// src/components/card/types.ts
export interface iServiceOrder {
  id: string;
  customername: string;
  purchasedate: string;
  customerphone: string;
  olhoesquerdodp: string;
  olhodireitodp: string;
  longeodesferico: string;
  longeodcilindro: string;
  longeodeixo: string;
  longeoeesferico: string;
  longeoecilindro: string;
  longeoeeixo: string;
  pertoodesferico: string;
  pertoodcilindro: string;
  pertoodeixo: string;
  pertooeesferico: string;
  pertooecilindro: string;
  pertooeeixo: string;
  addesquerdo: string;
  adddireito: string;
  tratamentoslentes: string;
  obs: string;
  createdat: string;
  updatedat: string;
}

export interface iServiceOrderPost {
  customername: string;
  purchasedate: string;
  customerphone: string;
  olhoesquerdodp: string;
  olhodireitodp: string;
  longeodesferico: string;
  longeodcilindro: string;
  longeodeixo: string;
  longeoeesferico: string;
  longeoecilindro: string;
  longeoeeixo: string;
  pertoodesferico: string;
  pertoodcilindro: string;
  pertoodeixo: string;
  pertooeesferico: string;
  pertooecilindro: string;
  pertooeeixo: string;
  addesquerdo: string;
  adddireito: string;
  tratamentoslentes: string;
  obs: string;
}